make state explicit
it's a lot of parameters but it seems much more clear to pass them
in rather than assuming they will be set on the chart.

get rid of the very clever continuation trick i just put in. :-|

also stop caching barWidth and xUnitCount, which wouldn't work
anymore (since we have to look at the past and present at once).

this works but a lot of tests are failing